requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth
playwright
python-dotenv
